Update Dockerfile
There seems to be a bug with the `--mount=type=cache` option that causes 
builds to fail unpredictably. 

So, for now, disabling the shared sccache and rely on the docker build cache instead...
This means builds take longer but should be more reliable. 

See [this issue](https://github.com/moby/buildkit/issues/1956).