Remove the sequential signing example
This example existed because we had a sequential prototype first before
we made the code distributed, and it didn't cost anything to keep it around.
Now it is obsolete (does not use FROST) and the keygen protocol is about
to get changed too. Also, it doesn't seem useful enough to try to
reproduce this for the new protocols.