feat: unified keygen and ceremony ids. (#1394)
* feat: unified keygen and ceremony ids.

Unifies ceremony id for keygen and signing in a single counter
in the vaults pallet.

The storage migration uses the greater of the two existing ids
as the new unified id.

* chore: rename CeremonyIdCounter -> CeremonyIdProvider in threhsold pallet config.

* feat: make ceremony id counter global

This was achieved by moving the storage item out of the
instantiable vaults pallet and into the validator pallet. This
is arguably not the most semantically meaningful location
to store this, but for now it seems the best option, rather
than adding whole new pallet.