fix: dont skip ceremony_ids:
Problem:

We were incrementing the ceremony id even when we hadn't enough
ceremony participants to make a request. This could result in skipping
a ceremony id on the CFE, but the CFE assumes incrementing ceremony
ids.

Solution:

Only increment ceremony id if we are going to request a ceremony. This
means a ceremony request *might* not be able to return a ceremony_id,
so it now returns an Option<CeremonyId>.

The main side effects of this change are:
- the (informational) events emitted now contain the attempt_count
instead of the ceremony_id (since there might not be one)
- we use request_id in place of ceremony_id for the threshold nomination
seed. This should be fine (in fact arguably more correct than the old
version).