fix: allow test coverage to run
This commit prevents a compiler issue with the code coverage test runner. The issue was a compiler overflow when evaluating some trait bounds related to the Grandpa pallet.

I wasn't able to solve the underlying problem - presumably it's something related to the internals of llvm-cov, and presumably it will be fixed in due course.

For now I have avoided the issue by using llvm-cov's cfg(coverage) compiler flag. I used this to conditionally compiler a more 'friendly' version of the offending trait, which is irrelevant to our code coverage. The 'normal' version of this remains active for all other tests and CI checks.