fix: ensure always unique IDs
The request ids could be reused, but submissions could last longer than the lifetime of the original request. Sometimes request_id (+ submission_id) was not unique. This was a particular problem, with `submission_status_futures` where an old submission could be confused as being for a newer request. And we could think a submission of the new request was in a block, when it wasn't, but actually it was just the old requests submission that was in the block.