feat: cherry pick (#4676)
* fix(github_actions/release_checks): update confusing runtime spec version check 🤦‍♂️ (#4672)

* fix: submission watcher could confuse/lose track of submissions (#4667)

* fix: ensure always unique IDs

The request ids could be reused, but submissions could last longer than the lifetime of the original request. Sometimes request_id (+ submission_id) was not unique. This was a particular problem, with `submission_status_futures` where an old submission could be confused as being for a newer request. And we could think a submission of the new request was in a block, when it wasn't, but actually it was just the old requests submission that was in the block.

* fix: allow submissions from different requests with matching nonce and submission_id

Basically previously we would lose track of some submissions under some specific unlikely situations. As we used a BTreeMap over submission_id which aren't globally unique (Only unique to a single request). This would previously cause the submission-watcher to miss a successful submission and retry or fail the request.

* refactor: simplify internal behaviour

The submission_status_future is no longer valid at this point, and we know it can be removed. By moving the remove outside the if, we ensure submission_status_futures is always inline with the submissions in submission_by_nonce, i.e. that both storage items have matching understanding of the pending submissions.

Note the `self.submissions_by_nonce.remove(nonce)` above this. Here the change is basically that everytime we remove i.e. `self.submissions_by_nonce.remove(nonce)` we should remove all the matching submissions from `submission_status_futures` so they are kept up to date.

**This was not a bug**

* fix: cf_pools_environment rpc encoding (#4674)

* chore: bump version to 1.3.1

---------

Co-authored-by: Assem <asem.hasna@gmail.com>
Co-authored-by: Alastair Holmes <42404303+AlastairHolmes@users.noreply.github.com>