Removed NumberOrHex from LP-API
Removed all usage of NumberOrHex from the LP inclusive all types that are relying on NumberOrHex. Removed all usage of NumberOrHex from the LP inclusive all types that are relying on NumberOrHex. Replaced it with a generic functions that takes a U256 and tries to parse it to the type of the generic argument.