Denote broadcast timeout in target chain blocks. (#5270)
* Denotate broadcast timeout in target chain blocks.

This change involves the following (in the broadcast pallet):
 - `BroadcastTimeout` is now specified in `ChainBlockNumberFor<T,I>`
 - `Timeouts` is now specified in `ChainBlockNumberFor<T,I>`
 - A new constant `SafeModeBlockMarginForTargetChain`, also in `ChainBlockNumberFor<T,I>`
   is introduced since the previously used `SafeModeBlockMargin` stays denominated
   in `BlockNumberFor<T>` for other purposes.
 - For convenience, introduction of block time constants for all target chains.
 - Update the migration introduced in #5258 since it hasn't been released yet.

* Cleanup, fix compilation, fix benchmark & mock.

* Run cargo fmt.

* Fix some tests which break after switching to external blocks.

* Fix rest of tests.

* Use new `then_execute_with_keep_context`.

* Test handling of multiple timeouts in same block.

With the current changes there might be multiple external blocks
being processed in the same statechain block. This commit adds a
test to make sure that all timed out validators are handled, even if
they were due in the past.

This commit also changes how the `ready_to_abort_broadcast` function
handles nominees. Previously, it assumed that the nominee is the first
entry of `MockEpochInfo::current_authorities`. This is only true for the
first broadcast that is mockingly aborted, and thus does not work in our
new test. Thus we update this function to extract the actual current nominee
from the `Timeouts` storage value instead.

* Fix call to wrong function name after rebase.

* Refactor tests to forward expiry target as part of the context instead of as mutable variable.

* Remove commented code.

* Change `Timeouts` into `StorageValue<Vec<...>>.`

It was previously a `StorageMap`, but checking for timeouts required
iteration over the whole map, and in such a usecase a vector inside a
value should be more performant.

* Apply suggestions from review.

* Apply more fixes from PR review:

 - Use constants instead of magic numbers.
 - Rename `SafeModeBlockMarginForTargetChain` to `SafeModeChainBlockMargin`.
 - Update comment describing usage of `Timeouts`.

* Apply last PR suggestions.

* Don't use a `BTreeSet` for timeouts, instead flatten into existing vector.

* Add migration for `Timeouts`.

We migrate from denoting timeouts by chainflip blocks to external
blocks. For existing timeouts in the storage we don't try to translate
them into the new format. Instead we simply set their new timeout value
to be "the current external block height + broadcast_timeout", thus
effectively resetting them.

* Update storage version numbers in runtime and in cf-broadcasts.

* Fix migration to allow for >= timeout.

* Use `drain` instead of manually removing entries.

* Properly use `T::ChainTracking` instead of `BlockHeightProvider`.

* Apply clippy suggestions.

* Improve migration test:

 - Test that old `Timeouts` is empty after migration.
 - Create some timeouts before running migration tests.