refactor: Use `EncodedAddress` for `cf_get_open_deposit_channels` (#5540)
* Use `EncodedAddress` for `cf_get_open_deposit_channels`.

This ensures that bitcoin addresses are returned as byte arrays instead of
using the internal encoding of Taproot etc.

* fix: restart deposit monitor on runtime upgrade.

Move code to start deposit monitor from temporary to permanent location.
This is because the same code should work in upgrades where we update from
on version of CFDM to another.

Unfortunately, this also adds 10 seconds waiting time to wait for CFDM to start.

* Better check for deposit monitor startup.

* chore: change polling frequency for deposit monitor startup.