fix: Correctly handle block_height=0 special case.
BW elections tracker now works in the special case where we want to create
elections starting from blockheight 0. Previously this was handled wrong
because we used `saturating_backward`. Now we changed the indexing to use `next_election`
instead of `highest_election`, allowing us to treat the 0 case in the same way as all others.