feat: CFE check for CF chain node connection (#6014)
When CFE is launched, it will try to check for CFE version - something
unique to CF nodes. This ensures Engine is connecting to a valid CF node,
not other Substrate based chains by mistake

If invalid CFE version is returned, the CFE exits early.

To reproduce the error and test the fix, replace `$SC_RPC_PORT` to `9947`
in `localnet/init/scripts/start-engine.sh`