feat: support localnet with static dev chainspec ⛓️ (#6047)
* chore: suporrt localnet with static dev chainspec ⛓️

* fix: pass CHAIN variable correctly 🐛

* fix: pass CHAIN variable thorugh export 🐛

* feat: adds chain input to workflows ✨

Adds a `chain` input to the `post_check.yml` workflow.

This allows specifying the chain environment (e.g., "dev", "testnet") for post-check runs.
Also, sets the chain in upgrade-test.yml and logs it in the start-all-nodes script.

* ci: disables upgrade-test job 🔇

Disables the upgrade-test job in the development CI workflow.

This change prevents the upgrade-test workflow from running
unnecessarily in the development environment.